pyproject.toml
[project]
name = "blocknote-py"
version = "0.2.0"
description = "🚀 BlockNote Python library - Convert BlockNote.js blocks to HTML, Markdown, PDF & JSON. Type-safe Pydantic models for Django, FastAPI, Flask backends. Rich text editor content processing made easy."
authors = [{ name = "Rohan Sharma Sitoula", email = "rohansharmasitoula@gmail.com" }]
maintainers = [{ name = "Rohan Sharma Sitoula", email = "rohansharmasitoula@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = [
    # Core BlockNote terms
    "blocknote", "blocknote-py", "blocknote-python", "blocknotejs", "block-note",
    # Framework integration
    "django", "fastapi", "flask", "pydantic", "python-backend",
    # Content formats
    "html", "markdown", "pdf", "json", "rich-text", "wysiwyg",
    # Use cases
    "text-editor", "content-management", "cms", "blog", "documentation",
    "document-processing", "content-conversion", "markup-converter",
    # Technical terms
    "server-side-rendering", "api", "rest-api", "web-development",
    "content-api", "headless-cms", "structured-content", "editor-backend"
]
classifiers = [
    # Development status
    "Development Status :: 4 - Beta",
    # Audience
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    # Topics - more specific and comprehensive
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Text Processing :: Filters",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Documentation",
    "Topic :: Communications :: Chat",
    # Framework specific
    "Framework :: Django",
    "Framework :: FastAPI",
    "Framework :: Flask",
    "Framework :: Pydantic",
    # Environment
    "Environment :: Web Environment",
    "Environment :: Console",
    # Python versions
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    # License and OS
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    # Additional metadata
    "Typing :: Typed",
    "Natural Language :: English",
]
dependencies = [
    "pydantic>=2.0.0",
    "markdown-it-py>=3.0.0",
]

[project.urls]
"Homepage" = "https://rohansharmasitoula.github.io/blocknote-py/"
"Documentation" = "https://rohansharmasitoula.github.io/blocknote-py/"
"Repository" = "https://github.com/rohansharmasitoula/blocknote-py"
"Source Code" = "https://github.com/rohansharmasitoula/blocknote-py"
"Bug Tracker" = "https://github.com/rohansharmasitoula/blocknote-py/issues"
"Issues" = "https://github.com/rohansharmasitoula/blocknote-py/issues"
"Changelog" = "https://rohansharmasitoula.github.io/blocknote-py/changelog/"
"Examples" = "https://rohansharmasitoula.github.io/blocknote-py/examples/"
"API Reference" = "https://rohansharmasitoula.github.io/blocknote-py/api/"
"PyPI" = "https://pypi.org/project/blocknote-py/"
"Funding" = "https://buymeacoffee.com/sitoularohansharma"
"Download" = "https://pypi.org/project/blocknote-py/#files"

[project.optional-dependencies]
pdf = [
    "weasyprint>=62.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]
all = [
    "blocknote-py[test,dev,docs,pdf]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.force-include]
"src" = "blocknote"

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "markdown_it.*",
    "tests.*",
    "*.__tests__.*",
]
ignore_missing_imports = true
ignore_errors = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=5.0.0,<6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.envs.default]
dependencies = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.hatch.envs.default.scripts]
lint = "uv run flake8 src"
format = "uv run black src && uv run isort src"
type-check = "uv run mypy src"
check-all = "uv run flake8 src && uv run black --check src && uv run isort --check-only src && uv run mypy src"
fix-all = "uv run black src && uv run isort src"

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__tests__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
